requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
PyJWT==2.8.0
cryptography==41.0.7
python-decouple==3.8
mysqlclient==2.2.0
redis==5.0.1
django-redis==5.4.0
djoser==2.3.1